feat(search-games): Add GameCard component and use in search-games page
This commit adds a new "searchGames" folder to the "components" directory, which contains the GameCard component used to display game information. The search-games page has also been updated to load the "games" props from getServerSideProps(), save the list of games in a useState array, and then map over this array to display the game cards using the GameCard component.

By separating out the GameCard component, we can reuse it in search-games page and keep our application and code more modular and maintainable.