refactor(search-games): Rename games props and useState constants
This commit refactors the "search-games.js" page by renaming the key of the props object returned by getServerSideProps() from "games" to "gamesData". The useState constants are also renamed from "[gamesList, setGamesList]" to "[games, setGames]". In addition, this commit automatically fixes a mistake in the mapping of the game list by using the useState array instead of the props object by renaming.