feat(utils): Add timestamp to date conversion function
This commit adds a new function in "utils/convertTimestamp.js" file that converts a timestamp object to a date string. This function will be used in the "search-games.js" page to compare release dates for sorting the game list.

This additional function improves the maintainability of the code by creating a reusable function for timestamp to date conversion, and enabling more efficient sorting of the game list.