fix(auth): increase interval time for token refreshing
Changed the interval time of the second useEffect hook in auth.js from 1 minute to 10 minutes. This was done to fix the "auth/user-token-expired" error that occurred when trying to update the user's password using the updatePassword() method of Firebase. Increasing the interval time to 10 minutes allows enough time for the updatePassword() method to complete and issue a new token, preventing the use of an old token and the resulting error.