fix(account): fix onAuthStateChanged unsubscribe issue
This commit fixes the "Error: Abort fetching component for route: '/signin'" issue caused by the onAuthStateChanged listener. The issue was caused because the listener was not being properly unsubscribed, which was leading to multiple executions of the listener and causing the error.

To fix the issue, the unsubscribe method was added to the useEffect hook to properly unsubscribe from the onAuthStateChanged listener.