feat(account): add support for Google provider re-authentication
Previously, only email and password provider was considered for re-authentication to handle updating user's password and deleting their accounts. Now, modified the code to consider all sign-in providers, including Google. Users will re-authenticate based on the provider they are signed in with, which (token.sign_in_provider) is generated in getServerSideProps() and passed as props to the page and its components.

Additionally, users who have signed up with Google provider can add email and password provider if they wish to.