feat(game, search): compute average rating and score
Compute the average rating and score for game(s) retrieved from the Firestore database in the getServerSideProps function of the search-games and game/[gameID] pages and include the calculated averages in the props returned to the client. If a game has no ratings or scores, the average is set to 0.

This ensures that the average rating and average score are accurate and calculated everytime whenever new reviews are added to the game.

Additionally, some small fixes / refactors are made\:
- Handle cases where there are no ratings or scores available for the game page.
- Set system requirements to null if they are not available in game page.
- Handle games data object if error occurs while fetching games collection for search page.