feat(reviews): add review component for reusability
The changes include adding a new Review component and refactoring the CriticReviews and UserReviews components to use the new Review component. The Review component contains the code that was previously in the CriticReviews and UserReviews components that displayed the reviews.

The new code is more reaable, modular and makes the component reusable in other parts of the code.

Some additional fixes/refactors are included as well:
- Remove getUserData() function and use its functionality directly (userReviews.js)
- Parse the score input's value to integer (criticReviews.js)
- Null check for conditionally rendering the critic reviews (criticReviews.js)