feat(layout): update navbar and footer components
This commit mainly refactors navbar and footer components to improve the overall design and user experience. Their layouts are restructed and styled with some addition of new features.

Modify the footer component to:
- use Link instead of Image for logo
- add social media links with icons
- add copyright section
- add styles based on mobile-first approach

Update the navbar component to:
- add a hamburger menu
- show navbar modal when the menu icon is clicked, otherwise, hide it
- use Link instead of Image for logo
- add react-icons to page navigation links
- add styles based on mobile-first approach

Additionally, update the global styles in styles/globals.css to include new styles for the footer and navbar components.