feat(next.config): add domain to image list
Add the domain "180dc.org" to the list of allowed domains for images in the next.config.js file. This change was made to ensure that images from the domain can be displayed on the website. Additionally, this fixes the user default image bug when visiting the user account page.