bug(_app): fix page loading behavior
The page loading behavior is modified to prevent it from making a complete load of any page.

Previously, the app temporarily removed a page from rendering during the loading phase and displayed the loading indicator in its place. Hence, it caused the page to make a complete reload with newly initialized states after the loading phase is completed. This behavior has been changed so that the page will remain unaltered with its states while the loading indicator is only displayed during the loading phase and disappears once the page has loaded.

Additionally, this commit adds a background blur effect to the loading screen to improve user experience.