feat: add loading spinner during asynchronous operation
This commit adds the use of the new loading spinner component and loading context to all the pages and their components. The useLoading function from the loading context is imported and used to set the loading state of the page. When the page needs to be in a loading state, such as when updating user profile, or deleting user profile, or loggin in or out, or posting reviews, the setIsPageLoading function is called to set the loading state to true, and then to false once the operation is complete. In short, the setIsPageLoading function is set to true before any asynchronous operation and set to false once the operation completes.

This provides users with a better visual experience while waiting for the application to complete its tasks.