feat(homepage): add links to buttons
This commit adds links to the "Explore" and "Join us" buttons in the hero section and the "Join for free" button in the value proposition section. Now, when users click on these buttons, they will be redirected to the corresponding pages. This improves the user experience by making it easier for them to navigate to the desired pages.