fix(account): fix page title and form submission
This commit fixes the errors of account.js page by resolving the page title and form submission bug. Before this commit, the page title did not update when user updated his display name. In addition, the form submission for deleting user account was getting cancelled every time and displayed the "Form submission canceled because the form is not connected" error. In this commit, both of the bugs are fixed.

Changes made:
- Use router.push() to render the page title with updated user's display name
- Add the type attribute of "button" to the cancel button in oder to avoid "form submission canceled" error