feat(account): add sign out button and functionality
This commit adds a sign out button to the account page, which allows the user to sign out of their account. Upon clicking the button, userSignOut() will be called. With the use of this function, the user will be signed out and a success or error toast message will be displayed accordingly. The loading state will be updated during the sign-out process. Additionally, this commit adds an additional styles (min height, and divider line for each component).