feat: update global styles and improve code readability
This commit updates the global styles to include new styles for HTML elements and utility classes in order to make it easier to style various elements throughout the website. Subsequently, the pages and their components are refactored to use the global styles, thus, making the components and subcomponents more visually appealing and easier to read.

Changes in globals.css:
- Update styles for h1, p, and button elements to ensure better consistency throughout the app
- Add new style rules for reusable components such as .heading, .primary-btn, .primary-btn--hover, .secondary-btn, .info-container, .rating-container, .empty-starts, .filled-stars, .modal-bg, and .modal-btn-container
- Rename some utility classes to improve consistency and readability such as .typing-input, .card-container, and.social-link, .filter-container, and .game-tab
- Define additional styles for filter UI using .filter-container, with child elements such as select and label

Changes in pages and components:
- Use, or replace class names with utility classes from globals.css
- Remove redundant classnames
- Add some concise ones for readability and consistency
- Modify the JSX syntax and change some class names
- Simplify the section's HTML structure
- Use semantic HTML elements where appropriate
- Change the text format and the layout of some elements

This commit will improve the readability of the pages and components making them more maintainable in the future.