feat(components): add reusable modal component
This commit adds a new reusable modal component that can be used across multiple components in the project. Subsequently, several components (deleteAccount, updatePassword, userProfile, and forgotPassword) are refactored to incorporate the new component. As a result, this commit ensures reusability of code, thus, preventing the use of separate modal elements.