fix(Image): add "sizes" and "priority" attributes
This commit fixes the "fill" warning message and improves page performance by adding "sizes" attributes to appropriate <Image> components in Next.js. Additionally, the "priority" attribute is added to the <Image> components for the Home and [gameID] pages, where the images are detected as the Largest Contentful Paint (LCP).

Modifications:
- Modify the SearchGames page and GameCard component to include an "index" prop

Additions:
- Add "sizes" attribute to Image component for GameCard, UserProfile, and Review components
- Add "priority" attribute to Image component for Home, and [gameID] pages
- Only add "priority" prop to the thumbnail image for the first GameCard component

Miscellaneous:
- Change order of attributes for Image component in Footer component for consistency and readability