style(search-games): adjust gradient and height
This commit includes changes to adjust the height of the gradient above the search games page header, as well as adding responsive values for different screen sizes. Specifically, the following changes are made:

- Change the height of the gradient to be 0.75rem on extra small screens, 1.25rem on small screens, and 1.75rem on meidum-to-larger screens.
- Remove the second gradient and adjusted the first gradient to fill the full height of the whole gradient.

These changes are made to fix and adjust the visual appearance of the transition between navbar and page header on different screen sizes.