fix(signin): update background image url
This commit updates the background image URL to an absolute path starting with a forward slash ("/community-controller.jpg") for the the quote section of the Sign in page. This fixes the issue where the image was not showing up on the production build due to an unknown URL scheme error.