refactor(package): update package name
This commit updates the package names in package.json and package-lock.json to reflect the project's new name, "gameon", which was recently implemented. The previous package name, "gameon-game-website", is no longer relevant, and this change will make the project easier to identify and manage.

This commit includes no new features, but is a necessary refactoring to keep the project up to date.